Put admin routes to a separate file and started skelton of user routes.
