+++ server/api/shop.js
@@ -128,47 +128,4 @@ router.post('/albums', requireToken, isAdmin, async (req, res, next) => {
 
 // PUT api/shop/order
 
-//DELETE api/shop/order
-
-//ADMIN PATHS
-router.post('/albums', requireToken, isAdmin, async (req, res, next) => {
-  try {
-    const { name, price, qty, releaseDate, label } = req.body;
-    const artistId = Math.floor(Math.random() * (100 - 1) + 1);
-    const product = await Product.create({
-      name,
-      price,
-      qty,
-      releaseDate,
-      label,
-      totalTrack: 0,
-      artistId,
-    });
-    res.send(product);
-  } catch (error) {
-    next(error);
-  }
-});
-
-//TODO: GET PRODUCT FROM DB AND UPDATE WITH NEW INFORMATION
-router.put('/albums/:id', requireToken, isAdmin, async (req, res, next) => {
-  try {
-    const { name, price, qty, releaseDate, label } = req.body;
-    const album = await Product.findByPk(req.params.id);
-    res.send(await album.update({ name, price, qty, releaseDate, label }));
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.delete('/albums/:id', requireToken, isAdmin, async (req, res, next) => {
-  try {
-    const product = await Product.findByPk(req.params.id);
-    await product.destroy();
-    res.send(product);
-  } catch (error) {
-    next(error);
-  }
-});
-
 module.exports = router;
