+++ src/components/admin/AdminDashboard.jsx
@@ -2,13 +2,22 @@ import React from "react";
 import DropdownActions from "./DropdownActions";
 import ProductsTable from "./ProductsTable";
 import UsersTable from "./UsersTable";
+import { useDispatch } from "react-redux";
+import { getUsers } from "../../reducers/adminReducer";
 
 const AdminDashboard = () => {
+  const dispatch = useDispatch();
   const [page, setPage] = React.useState(false);
+
+  React.useEffect(() => {
+    dispatch(getUsers());
+    console.log(`what is the page?`, page);
+  }, []);
+
   return (
     <div className="adminDashboard">
       <div className="options">
-        <DropdownActions />
+        <DropdownActions page={page} />
         <button
           onClick={() => {
             setPage(!page);
